// src/app/main-appli-module/common-module/models/scale-poster.enum.ts
export enum ScalePoster {
    SCALE_100h = '100h',
    SCALE_300h = '300h',
    SCALE_350h = '350h',
    SCALE_600h = '600h',
    SCALE_900h = '900h',
    SCALE_1400h = '1400h',
    SCALE_1920h = '1920h',

    SCALE_100w = '100w',
    SCALE_300w = '300w',
    SCALE_350w = '350w',
    SCALE_600w = '600w',
    SCALE_900w = '900w',
    SCALE_1400w = '1400w',
    SCALE_1920w = '1920w',

    SCALE_200w = '200w',
    SCALE_500w = '500w',
    SCALE_700w = '700w',

    SCALE_ORIGINAL = 'SCALE_ORIGINAL'
}